Redirect the user based on the login status.
